finalizacion crud con sweet alert

+++ ezpeletaNetCore8/wwwroot/js/ejercicioFisico.js
@@ -17,11 +17,31 @@ function listadoEjerciciosFisicos(){
         dataType: 'json',
         success: function(ejerciciosFisicos){
             $.each(ejerciciosFisicos, function(index, ejercicio){
+
+                // Formatear el objeto Date al formato deseado
+                var fechaInicioFormateada = new Date(ejercicio.inicio).toLocaleDateString('es-ES', {
+                    day: '2-digit',
+                    month: '2-digit',
+                    year: 'numeric',
+                    hour: '2-digit',
+                    minute: '2-digit'
+                });
+
+                var fechaFinFormateada = new Date(ejercicio.fin).toLocaleDateString('es-ES', {
+                    day: '2-digit',
+                    month: '2-digit',
+                    year: 'numeric',
+                    hour: '2-digit',
+                    minute: '2-digit'
+                });
+
                 $('#tbody-ejerciciosFisicos').append(`
                     <tr>
                         <td class="tbody">${ejercicio.nombreTipoEjercicio}</td>
-                        <td class="tbody">${ejercicio.inicio}</td>
-                        <td class="tbody">${ejercicio.fin}</td>
+                        <td class="tbody">${fechaInicioFormateada} hs.</td>
+                        <td class="tbody">${fechaFinFormateada} hs.</td>
+                        <td class="tbody">${ejercicio.estadoEmocionalInicio}</td>
+                        <td class="tbody">${ejercicio.estadoEmocionalFin}</td>
                         <td class="text-center">
                             <button type="button" class="btn btn-success mb-2" onclick="abrirModalEditar(${ejercicio.ejercicioFisicoID})">
                                 Editar
@@ -37,14 +57,18 @@ function listadoEjerciciosFisicos(){
             })
         },
         error: function(hxr, status){
-            alert('Ocurrió un error a la hora de mostrar el listado.')
+            Swal.fire({
+                title: 'Ups, existe un inconveniente:',
+                text: 'Ocurrió un problema a la hora de mostrar el listado.',
+                icon: 'warning',
+                confirmButtonText: 'Volver a intentarlo'
+            });
         }
     })
 }
 
 
 function limpiarModal(){
-
     document.getElementById("TipoEjercicioID").value = 0;
     document.getElementById("FechaInicio").value = "";
     document.getElementById("EstadoEmocionalInicio").value = 0;
@@ -76,22 +100,45 @@ function guardarRegistro(){
             listadoEjerciciosFisicos();
         },
         error: function(hxr,status){
-            alert('Ocurrió un error al almacear el nuevo ejercicio físico.');
+            Swal.fire({
+                title: 'Ups, existe un inconveniente:',
+                text: 'Ocurrió un error al almacear el nuevo ejercicio físico.',
+                icon: 'warning',
+                confirmButtonText: 'Volver a intentarlo'
+            });
         }
     })
 }
 
 function eliminarRegistro(ejercicioFisicoID){
-    $.ajax({
-        url: '../../EjercicioFisico/EliminarEjercicioFisico',
-        data: { ejercicioFisicoID : ejercicioFisicoID },
-        type: 'DELETE',
-        dataType: 'json',
-        success: function(result){
-            listadoEjerciciosFisicos();
-        },
-        error: function(kxr,status){
-            alert('Disculpe, ocurrió un error al intentar eliminar el ejercicio físico.');
+    Swal.fire({
+        title: "¿Desea eliminar el ejercicio físico?",
+        icon: "warning",
+        showCancelButton: true,
+        confirmButtonColor: "#3085d6",
+        cancelButtonColor: "#d33",
+        confirmButtonText: "Confirmar",
+        cancelButtonText: "Cancelar"
+
+    }).then((result) =>{
+        if (result.isConfirmed) {
+            $.ajax({
+                url: '../../EjercicioFisico/EliminarEjercicioFisico',
+                data: { ejercicioFisicoID : ejercicioFisicoID },
+                type: 'DELETE',
+                dataType: 'json',
+                success: function(result){
+                    listadoEjerciciosFisicos();
+                },
+                error: function(kxr,status){
+                    Swal.fire({
+                        title: 'Ups, existe un inconveniente:',
+                        text: 'Disculpe, ocurrió un error al intentar eliminar el ejercicio físico.',
+                        icon: 'warning',
+                        confirmButtonText: 'Volver a intentarlo'
+                    });
+                }
+            })
         }
     })
 }
@@ -113,8 +160,8 @@ function abrirModalEditar(ejercicioFisicoID){
             document.getElementById("FechaInicio").value = ejercicioAEditar.inicio;
             document.getElementById("FechaFin").value = ejercicioAEditar.fin;
             document.getElementById("Observaciones").value = ejercicioAEditar.observaciones;
-            document.getElementById("EstadoEmocionalInicio").value = ejercicioAEditar.estadoEmocionalInicio;
-            document.getElementById("EstadoEmocionalFin").value = ejercicioAEditar.estadoEmocionalFin;
+            document.getElementById("EstadoEmocionalInicio").value = ejercicioAEditar.estadoEmocionalInicioInt;
+            document.getElementById("EstadoEmocionalFin").value = ejercicioAEditar.estadoEmocionalFinInt;
             $('#modalEjercicioFisico').modal("show");
 
         }
